moved signallist, adclist, slowadc, slowadclist to Ctb
